+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/Patchable.java
@@ -28,7 +28,7 @@ public interface Patchable<T> {
    * WARNING: This may overwrite concurrent changes (between when you obtained your item and the current state) in an unexpected way.
    * Consider using edit instead.
    *
-   * @param item item to be patched with patched values
+   * @param item to be patched with patched values
    * @return returns deserialized version of api server response
    */
   default T patch(T item) {
@@ -41,14 +41,14 @@ public interface Patchable<T> {
    * <ul>
    * <li>{@link PatchType#JSON} - will create a JSON patch against the current item.
    * WARNING: This may overwrite concurrent changes (between when you obtained your item and the current state) in an unexpected way.
-   * Consider using edit instead.
+   * Consider using edit instead.  The resourceVersion and other fields of the item may be modified by this call.
    * <li>{@link PatchType#JSON_MERGE} - will send the serialization of the item as a JSON MERGE patch.
    * Set the resourceVersion to null to prevent optimistic locking.
    * <li>{@link PatchType#STRATEGIC_MERGE} - will send the serialization of the item as a STRATEGIC MERGE patch.
    * Set the resourceVersion to null to prevent optimistic locking.
    * </ul>
    *
-   * @param item item to be patched with patched values
+   * @param item to be patched with patched values
    * @param patchContext {@link PatchContext} for patch request
    * @return returns deserialized version of api server response
    */