+++ kubernetes-model-generator/kubernetes-model-core/src/main/java/io/fabric8/kubernetes/internal/KubernetesDeserializer.java
@@ -183,7 +183,9 @@ public class KubernetesDeserializer extends JsonDeserializer<KubernetesResource>
                 "io.fabric8.openshift.api.model.operatorhub.manifests.",
                 "io.fabric8.openshift.api.model.operatorhub.v1.",
                 "io.fabric8.openshift.api.model.operatorhub.v1alpha1.",
+                "io.fabric8.openshift.api.model.operatorhub.lifecyclemanager.v1.",
                 "io.fabric8.openshift.api.model.machineconfig.v1.",
+                "io.fabric8.openshift.api.model.tuned.v1.",
                 "io.fabric8.kubernetes.api.model.extensions."
         };